+++ lizard_measure/tests.py
@@ -1,8 +1,10 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
 
 from django.test import TestCase
-
 from lizard_measure.models import Measure
+
+import logging
+logger = logging.getLogger(__name__)
 
 
 class ModelTest(TestCase):
@@ -12,5 +14,5 @@ class ModelTest(TestCase):
         pass
 
     def test_cost_attribute(self):
-        m = Measure.objects.get('TestMeasure1')
-        self.assertEquals('m.cost', 500)
+        m = Measure.objects.get(title='TestMeasure1')
+        self.assertEquals(m.total_costs, 500)